+++ front/src/components/ConsultaRapidaComponent/PacienteData.js
@@ -1,34 +1,11 @@
 import React from 'react';
 
-const DATA_TEMPORAL = [
-    {
-        title:'Nombre',
-        data:'Juana Perez'
-    },
-    {
-        title:'Teléfono',
-        data:'000-000-0000'
-    },
-    {
-        title:'Correo',
-        data:'juana.perez@gmail.com'
-    },
-    {
-        title:'Tipo de cita',
-        data:'Terapia inicial'
-    },
-    {
-        title:'Este es mi mensaje',
-        data:'Hola solo quiero decir que...'
-    }
-]
-
-export default function PacienteData(){
+export default function PacienteData(props){
     return <div>
         <h5>Estos son tus datos:</h5>
-        {DATA_TEMPORAL.map((index,key)=><span key={key}>
+        {props.DATATEMPORAL.map((index,key)=><span key={key}>
             <h4>{index.title}</h4>
-            <h5>{index.data}</h5>
+            {typeof index.data === 'object' ? index.data.map(index => <h5>{index}</h5>) : <h5>{index.data}</h5>}
         </span>)}
     </div>
 }
