+++ front/src/scss/EnviarResultados.scss
@@ -1,0 +1,123 @@
+.class-EnviarResultados-step1 {
+    padding: 0px 5%;
+    display: grid;
+    grid-template-columns: 32.5% 32.5% 32.5%;
+    grid-gap: 20px;
+    height: calc(100vh - 107px);
+    align-content: center;
+
+    > div:nth-child(2) {
+        display: flex;
+        flex-direction: column;
+        gap: 20px;
+    }
+
+    > span:first-child {
+        display: flex;
+        flex-direction: column;
+        gap: 40px;
+
+        .MuiFormControl-root.css-1nrlq1o-MuiFormControl-root {
+            gap: 20px;
+        }
+    }
+
+    > div:nth-child(3) {
+        justify-self: end;
+    }
+
+    > div:last-child {
+        grid-column: 3;
+        justify-self: end;
+    }
+
+    > div:nth-child(3) {
+        position: relative;
+        width: 100%;
+        height: 100%;
+    }
+}
+
+.class-EnviarResultados-step2 {
+    padding: 0px 5%;
+    display: grid;
+    grid-template-columns: 25% 25% auto;
+    grid-gap: 20px;
+    height: calc(100vh - 107px);
+    align-content: center;
+
+    h4 {
+        font-weight: bold;
+    }
+
+    > div:nth-child(2) {
+        display: grid;
+    }
+
+    > div:last-child {
+        grid-column: 3;
+        justify-content: right;
+    }
+
+    form {
+        display: grid;
+        grid-template-columns: calc(50% - 10px) calc(50% - 10px);
+        grid-template-rows: repeat(5,56px);
+        gap: 20px;
+
+        > h5,
+        > div:nth-child(4),
+        > div:nth-child(5)  {
+            grid-column: 1 / span 2;
+        }
+
+        > .css-xn6ruc>:not(style),
+        > div,
+        > div:nth-child(2) > div {
+            width: auto;
+            margin: 0;
+        }
+    }
+}
+
+
+.class-UploadButtons {
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    max-height: 507px;
+    overflow-y: auto;
+
+    .class-cardUploaded {
+        background: #efefef;
+        border-radius: 10px;
+        display: flex;
+        gap: 20px;
+        align-items: center;
+        padding: 20px;
+        color: #c4c4c4;
+    }
+
+    button {
+        background: transparent;
+        padding: 20px;
+        width: 100%;
+        border: 3px solid #c4c4c4;
+        border-style: dashed;
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        color: #1976d2;
+
+        h4 {
+            color: #1976d2;
+            text-decoration: underline;
+        }
+    }
+
+    button.class-cardUpload {
+        flex-direction: row !important;
+        display: flex;
+    }
+}
