+++ front/src/scss/GlobalStyle.scss
@@ -7,4 +7,5 @@
 @import "./filterComponent.scss";
 @import "./cardConsulta.scss";
 @import "./EditOrSeeDetailsComponent.scss";
-@import "./AgendarCita.scss"
+@import "./AgendarCita.scss";
+@import "./EnviarResultados.scss";
